GUI/requirements.txt
gTTS==2.5.3
IMDbPY==2022.7.9
keyboard==0.13.5
Kivy

Kivy_examples==2.3.0
PyAutoGUI==0.9.54
pydub==0.25.1
python-decouple==3.8
python-dotenv==1.0.1
pyttsx3==2.90
pywhatkit==5.4
Requests==2.32.3
sounddevice==0.4.6
SpeechRecognition==3.10.0
wikipedia==1.4.0
wolframalpha==5.0.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.2.0
altair==5.4.1
annotated-types==0.6.0
anyio==4.6.0
attrs==23.1.0
blinker==1.8.2
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
distlib==0.3.6
filelock==3.12.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage
google-api-core==2.19.1
google-api-python-client==2.147.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.2
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain-core==0.3.9
langchain-google-genai==2.0.0
langsmith==0.1.131
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.4
narwhals==1.9.1
numpy==2.1.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip==24.2
platformdirs==3.5.0
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
# PyAudio==0.2.11 (Install manually if needed)

pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.7.2
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.4.4
streamlit==1.39.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.23.0
watchdog==5.0.3
wheel==0.44.0
win-inet-pton==1.1.0
yarl==1.11.0
